+++ Source/Gunner/Private/Tests/GunnerMotionProbe.h
@@ -2,6 +2,7 @@
 
 #include "CoreMinimal.h"
 #include "GameFramework/Actor.h"
+#include "Tests/GunnerProbeInputGuard.h"
 #include "GunnerMotionProbe.generated.h"
 
 class AGunnerCharacter;
@@ -20,6 +21,7 @@ public:
     virtual void Tick(float DeltaSeconds) override;
     virtual void EndPlay(const EEndPlayReason::Type Reason) override;
 private:
+    FGunnerProbeInputGuard InputGuard;
     UPROPERTY() TObjectPtr<AGunnerCharacter> Character;
     UPROPERTY() TObjectPtr<APlayerController> Player;
     UPROPERTY() TObjectPtr<AGunnerTarget> Target;